write custom header identifying the format to h5mu files
This makes use of the user-block feature of HDF5. Also save this
information as top-level attributes for the file and for each modality
group. This might be useful if an individual modality is replaced by
another muon version while leaving the rest intact, or if someone wants
to save a MuData as a group in a file containing other stuff